Support signing and verification using ECDSA
Introduce the methods `crypto.subtle.sign()` and
`crypto.subtle.verify()` to create and verify ECDSA signatures.

Currently, only the combination of the algorithm `ECDSAinDERFormat` and
`SHA-256` hash is supported.

The algorithm `ECDSAinDERFormat` is similar to the `ECDSA` algorithm
used in `SubtleCrypto`, with the distinction that signature is encoded
in the DER format, diverging from the IEEE-P1363 format used in
`SubtleCrypto`.