Signing and key derivation: support configuring auth prompt texts
For keys that require authentication, a prompt is shown to the user on
Android to confirm the operation. The texts were previously hard-coded.
This change allows to configure them via an options object passed to
`subtle.sign()`, `subtle.deriveKey()` and `subtle.deriveBits()`. When
the key is not protected by a user authentication requirement, the
authentication options are ignored.

Prefer using an object over two separate parameters to allow adding
more Tabris.js proprietary options in the future.